+++ index.html
@@ -31,6 +31,7 @@
 </head>
 
 <body>
+  <img id="screenshot-img" src="ui_assets/3DStreet-Viewer-Start-Editor.svg" alt="Invisible Image" style="display:none;">
   <!-- loading animation start -->
   <div class="loader__wrapper">
     <div class="loader">
@@ -47,7 +48,6 @@
       <div class="road">Loading 3DStreet</div>
     </div>
   </div>
-
   <!-- viewer ui start -->
   <div class="viewer-header-wrapper">
     <button class="viewer-logo-start-editor-button" onclick="startEditor()">
