+++ src/editor/components/modals/ScreenshotModal/ScreenshotModal.component.jsx
@@ -1,127 +1,17 @@
 import { useEffect, useState } from 'react';
 import styles from './ScreenshotModal.module.scss';
-import {
-  collection,
-  doc,
-  getDoc,
-  serverTimestamp,
-  updateDoc
-} from 'firebase/firestore';
-
 import { signIn } from '../../../api';
-import { getDownloadURL, ref, uploadBytes } from 'firebase/storage';
 import PropTypes from 'prop-types';
 import { useAuthContext } from '../../../contexts';
 import { Copy32Icon, Save24Icon } from '../../../icons';
-import { db, storage } from '../../../services/firebase';
 import { Button, Dropdown, Input } from '../../components';
 import Toolbar from '../../scenegraph/Toolbar';
 import Modal from '../Modal.jsx';
 import posthog from 'posthog-js';
 import { saveBlob } from '../../../lib/utils';
 import Events from '../../../lib/Events';
-
+import { uploadThumbnailImage, saveScreenshot } from '../../../api/scene';
 // import { loginHandler } from '../SignInModal';
-
-export const uploadThumbnailImage = async (uploadedFirstTime) => {
-  try {
-    saveScreenshot('img');
-
-    await new Promise((resolve) => setTimeout(resolve, 1000));
-    const screentockImgElement = document.getElementById(
-      'screentock-destination'
-    );
-
-    // Get the original image dimensions
-    const originalWidth = screentockImgElement.naturalWidth;
-    const originalHeight = screentockImgElement.naturalHeight;
-
-    // Define the target dimensions
-    const targetWidth = 320;
-    const targetHeight = 240;
-
-    // Calculate the scale factors
-    const scaleX = targetWidth / originalWidth;
-    const scaleY = targetHeight / originalHeight;
-
-    // Use the larger scale factor to fill the entire space
-    const scale = Math.max(scaleX, scaleY);
-
-    // Calculate the new dimensions
-    const newWidth = originalWidth * scale;
-    const newHeight = originalHeight * scale;
-
-    const resizedCanvas = document.createElement('canvas');
-    resizedCanvas.width = targetWidth;
-    resizedCanvas.height = targetHeight;
-    const context = resizedCanvas.getContext('2d');
-
-    // Calculate the position to center the image
-    const posX = (targetWidth - newWidth) / 2;
-    const posY = (targetHeight - newHeight) / 2;
-
-    // Draw the image on the canvas with the new dimensions and position
-    context.drawImage(screentockImgElement, posX, posY, newWidth, newHeight);
-    // Rest of the code...
-    const thumbnailDataUrl = resizedCanvas.toDataURL('image/jpeg', 0.5);
-    const blobFile = await fetch(thumbnailDataUrl).then((res) => res.blob());
-
-    const sceneDocId = STREET.utils.getCurrentSceneId();
-
-    const thumbnailRef = ref(storage, `scenes/${sceneDocId}/files/preview.jpg`);
-
-    const uploadedImg = await uploadBytes(thumbnailRef, blobFile);
-
-    const downloadURL = await getDownloadURL(uploadedImg.ref);
-    const userScenesRef = collection(db, 'scenes');
-    const sceneDocRef = doc(userScenesRef, sceneDocId);
-    const sceneSnapshot = await getDoc(sceneDocRef);
-    if (sceneSnapshot.exists()) {
-      await updateDoc(sceneDocRef, {
-        imagePath: downloadURL,
-        updateTimestamp: serverTimestamp()
-      });
-      console.log('Firebase updateDoc fired');
-    } else {
-      throw new Error('No existing sceneSnapshot exists.');
-    }
-
-    console.log('Thumbnail uploaded and Firestore updated successfully.');
-    uploadedFirstTime &&
-      STREET.notify.successMessage(
-        'Scene thumbnail updated in 3DStreet Cloud.'
-      );
-  } catch (error) {
-    console.error('Error capturing screenshot and updating Firestore:', error);
-    let errorMessage = `Error updating scene thumbnail: ${error}`;
-    if (error.code === 'storage/unauthorized') {
-      errorMessage =
-        'Error updating scene thumbnail: only the scene author may change the scene thumbnail. Save this scene as your own to change the thumbnail.';
-    }
-    STREET.notify.errorMessage(errorMessage);
-  }
-};
-
-const saveScreenshot = async (value) => {
-  const screenshotEl = document.getElementById('screenshot');
-  screenshotEl.play();
-
-  if (value === 'img') {
-    screenshotEl.setAttribute(
-      'screentock',
-      'imgElementSelector',
-      '#screentock-destination'
-    );
-  }
-
-  posthog.capture('screenshot_taken', {
-    type: value,
-    scene_id: STREET.utils.getCurrentSceneId()
-  });
-
-  screenshotEl.setAttribute('screentock', 'type', value);
-  screenshotEl.setAttribute('screentock', 'takeScreenshot', true);
-};
 
 const filterHelpers = (scene, visible) => {
   scene.traverse((o) => {
